January/requirements.txt
chromadb>=0.4.0,<0.5.0
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas
requests
streamlit
pydantic-settings>=2.0.0
pydantic>=2.0.0,<3.0.0
scikit-learn
tensorflow>=2.15.0
seaborn
jupyter
transformers
tensorflow-hub
matplotlib>=3.4.0

# langchain
langchain
langchain-aws
langchain-community
langchain-text-splitters
boto3
python-dotenv
langgraph
langchain-openai
langsmith
langgraph-checkpoint-postgres
psycopg
ipykernel